+++ package.json
@@ -19,11 +19,13 @@
     "classnames": "^2.5.1",
     "cli": "^1.0.1",
     "clsx": "^2.1.1",
+    "dotenv": "^16.4.7",
     "fs": "^0.0.1-security",
     "lucide-react": "^0.459.0",
     "next": "^15.0.4",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
+    "resend": "^4.0.1",
     "shadcn": "^1.0.0",
     "shadcn-ui": "^0.2.3",
     "stripe": "^17.4.0",
